move *.browser and *.node to browser/* and node/*
...this should leave the lib folder a little bit less cluttered.
