+++ lib/Key.js
@@ -1,5 +1,5 @@
 if (process.versions) {
-  module.exports = require('./Key.node');
+  module.exports = require('./node/Key');
   return;
 }
-module.exports = require('./Key.browser');
+module.exports = require('./browser/Key');
