+++ lib/Point.js
@@ -1,5 +1,5 @@
 if (process.versions) {
-  module.exports = require('./Point.node');
+  module.exports = require('./node/Point');
   return;
 }
-module.exports = require('./Point.browser');
+module.exports = require('./browser/Point');
