+++ lib/browser/Key.js
@@ -1,4 +1,4 @@
-var ECKey = require('../browser/vendor-bundle.js').ECKey;
+var ECKey = require('../../browser/vendor-bundle.js').ECKey;
 var buffertools = require('buffertools');
 
 var Key = function() {
