+++ lib/browser/Point.js
@@ -4,11 +4,11 @@ var imports = require('soop').imports();
 var Key = imports.Key || require('./Key'); 
 var bignum = imports.bignum || require('bignum');
 var assert = require('assert');
-var ECKey = require('../browser/vendor-bundle.js').ECKey;
-var ECPointFp = require('../browser/vendor-bundle.js').ECPointFp;
-var ECFieldElementFp = require('../browser/vendor-bundle.js').ECFieldElementFp;
-var getSECCurveByName = require('../browser/vendor-bundle.js').getSECCurveByName;
-var BigInteger = require('../browser/vendor-bundle.js').BigInteger;
+var ECKey = require('../../browser/vendor-bundle.js').ECKey;
+var ECPointFp = require('../../browser/vendor-bundle.js').ECPointFp;
+var ECFieldElementFp = require('../../browser/vendor-bundle.js').ECFieldElementFp;
+var getSECCurveByName = require('../../browser/vendor-bundle.js').getSECCurveByName;
+var BigInteger = require('../../browser/vendor-bundle.js').BigInteger;
 var should = require('chai').should();
 
 //a point on the secp256k1 curve
