+++ lib/node/Point.js
@@ -1,7 +1,7 @@
 "use strict";
 
 var imports = require('soop').imports();
-var Key = imports.Key || require('./Key'); 
+var Key = imports.Key || require('../Key'); 
 var bignum = imports.bignum || require('bignum');
 var assert = require('assert');
